Refactor entity class to remove required modifier from Id property and update namespaces for consistency
