+++ benchmark/Audit/Wangkanai.Audit.Benchmark.csproj
@@ -1,7 +1,7 @@
 ﻿<Project Sdk="Microsoft.NET.Sdk">
 
    <ItemGroup>
-      <ProjectReference Include="..\..\src\Audit\Wangkanai.Audit.csproj"/>
+      <ProjectReference Include="..\..\src\Audit\Wangkanai.Audit.csproj" />
    </ItemGroup>
 
 </Project>
