+++ benchmark/Domain/Wangkanai.Domain.Benchmark.csproj
@@ -1,7 +1,7 @@
 <Project Sdk="Microsoft.NET.Sdk">
 
    <ItemGroup>
-      <ProjectReference Include="..\..\src\Domain\Wangkanai.Domain.csproj"/>
+      <ProjectReference Include="..\..\src\Domain\Wangkanai.Domain.csproj" />
    </ItemGroup>
 
 </Project>
