+++ benchmark/EntityFramework/Wangkanai.EntityFramework.Benchmark.csproj
@@ -1,7 +1,7 @@
 <Project Sdk="Microsoft.NET.Sdk">
 
    <ItemGroup>
-      <ProjectReference Include="..\..\src\EntityFramework\Wangkanai.EntityFramework.csproj"/>
+      <ProjectReference Include="..\..\src\EntityFramework\Wangkanai.EntityFramework.csproj" />
    </ItemGroup>
 
 </Project>
