+++ src/Audit/Wangkanai.Audit.csproj
@@ -5,12 +5,12 @@
       <PackageTags>wangkanai;audit;ddd;dotnet</PackageTags>
    </PropertyGroup>
    <ItemGroup>
-      <PackageReference Include="Microsoft.EntityFrameworkCore"/>
-      <PackageReference Include="Microsoft.EntityFrameworkCore.Relational"/>
-      <PackageReference Include="Microsoft.Extensions.Identity.Stores"/>
+      <PackageReference Include="Microsoft.EntityFrameworkCore" />
+      <PackageReference Include="Microsoft.EntityFrameworkCore.Relational" />
+      <PackageReference Include="Microsoft.Extensions.Identity.Stores" />
    </ItemGroup>
    <ItemGroup>
-      <ProjectReference Include="..\Domain\Wangkanai.Domain.csproj"/>
+      <ProjectReference Include="..\Domain\Wangkanai.Domain.csproj" />
    </ItemGroup>
 
 </Project>
