+++ src/Domain/Entity.cs
@@ -10,7 +10,7 @@ namespace Wangkanai.Domain;
 public abstract class Entity<T> : IEntity<T> where T : IEquatable<T>, IComparable<T>
 {
    /// <summary>Gets or sets the unique identifier for the entity. This property is used to uniquely identify an instance of the entity within the domain. The type of the identifier is defined by the generic type parameter of the entity.</summary>
-   public required T Id { get; set; }
+   public T Id { get; set; }
 
    /// <summary>Determines whether the entity is transient, meaning it has not been assigned a valid identifier. An entity is considered transient if its identifier equals the default value for its type.</summary>
    /// <returns>true if the entity is transient; otherwise, false.</returns>
