+++ src/Domain/Wangkanai.Domain.csproj
@@ -5,10 +5,10 @@
       <PackageTags>wangkanai;domain;ddd;audit;entityframework;dotnet</PackageTags>
    </PropertyGroup>
    <ItemGroup>
-      <PackageReference Include="Microsoft.EntityFrameworkCore"/>
-      <PackageReference Include="Microsoft.EntityFrameworkCore.Relational"/>
-      <PackageReference Include="Microsoft.Extensions.Identity.Stores"/>
-      <PackageReference Include="Wangkanai.System"/>
-      <PackageReference Include="Wangkanai.Validation"/>
+      <PackageReference Include="Microsoft.EntityFrameworkCore" />
+      <PackageReference Include="Microsoft.EntityFrameworkCore.Relational" />
+      <PackageReference Include="Microsoft.Extensions.Identity.Stores" />
+      <PackageReference Include="Wangkanai.System" />
+      <PackageReference Include="Wangkanai.Validation" />
    </ItemGroup>
 </Project>
