+++ src/EntityFramework/Wangkanai.EntityFramework.csproj
@@ -7,15 +7,15 @@
    </PropertyGroup>
 
    <ItemGroup>
-      <FrameworkReference Include="Microsoft.AspNetCore.App"/>
+      <FrameworkReference Include="Microsoft.AspNetCore.App" />
    </ItemGroup>
 
    <ItemGroup>
-      <PackageReference Include="Microsoft.EntityFrameworkCore.SqlServer"/>
+      <PackageReference Include="Microsoft.EntityFrameworkCore.SqlServer" />
    </ItemGroup>
 
    <ItemGroup>
-      <ProjectReference Include="..\Audit\Wangkanai.Audit.csproj"/>
+      <ProjectReference Include="..\Audit\Wangkanai.Audit.csproj" />
    </ItemGroup>
 
 </Project>
