+++ tests/Audit/Wangkanai.Audit.Tests.csproj
@@ -5,7 +5,7 @@
    </PropertyGroup>
 
    <ItemGroup>
-      <ProjectReference Include="..\..\src\Audit\Wangkanai.Audit.csproj"/>
+      <ProjectReference Include="..\..\src\Audit\Wangkanai.Audit.csproj" />
    </ItemGroup>
 
 </Project>
