+++ tests/Domain/Wangkanai.Domain.Tests.csproj
@@ -1,5 +1,8 @@
 <Project Sdk="Microsoft.NET.Sdk">
+   <PropertyGroup>
+      <RootNamespace>Wangkanai.Domain</RootNamespace>
+   </PropertyGroup>
    <ItemGroup>
-      <ProjectReference Include="..\..\src\Domain\Wangkanai.Domain.csproj"/>
+      <ProjectReference Include="..\..\src\Domain\Wangkanai.Domain.csproj" />
    </ItemGroup>
 </Project>
